consolidate main packages
