+++ examples/configs/collector-config-example.yaml
